+++ ext/libbpf_ruby/libbpf_ruby.c
@@ -303,6 +303,21 @@ static VALUE rb_cProgram_attach_cgroup(VALUE self, VALUE cgroup) {
   return libbpf_ruby_link_wrap(self, link);
 }
 
+static VALUE rb_cProgram_attach_perf_event(VALUE self, VALUE perf_event) {
+  struct bpf_program *program = libbpf_ruby_program_bpf(self);
+  int perf_fd = dup(rb_io_descriptor(perf_event));
+  if (perf_fd < 0) {
+    rb_raise(rb_eRuntimeError, "dup failed: %s", strerror(errno));
+  }
+  struct bpf_link *link = bpf_program__attach_perf_event(program, perf_fd);
+  long err = libbpf_get_error(link);
+  if (err) {
+    close(perf_fd);
+    rb_raise(rb_eRuntimeError, "bpf_program__attach_perf_event failed: %s", strerror(-err));
+  }
+  return libbpf_ruby_link_wrap(self, link);
+}
+
 static VALUE rb_cLink_fd(VALUE self) {
   libbpf_ruby_link_t *libbpf_ruby_link;
   TypedData_Get_Struct(self, libbpf_ruby_link_t, &libbpf_ruby_link_type, libbpf_ruby_link);
@@ -395,6 +410,7 @@ RUBY_FUNC_EXPORTED void Init_libbpf_ruby(void) {
   rb_define_method(rb_cLibBPFRubyProgram, "attach_tracepoint", rb_cProgram_attach_tracepoint, 2);
   rb_define_method(rb_cLibBPFRubyProgram, "attach_raw_tracepoint", rb_cProgram_attach_raw_tracepoint, 1);
   rb_define_method(rb_cLibBPFRubyProgram, "attach_cgroup", rb_cProgram_attach_cgroup, 1);
+  rb_define_method(rb_cLibBPFRubyProgram, "attach_perf_event", rb_cProgram_attach_perf_event, 1);
 
   rb_undef_alloc_func(rb_cLibBPFRubyLink);
   rb_define_method(rb_cLibBPFRubyLink, "fd", rb_cLink_fd, 0);